Common ViewPolicy
