+++ portals/publisher/src/main/webapp/source/src/app/components/GlobalPolicies/Policies/ViewPolicy.tsx
@@ -17,87 +17,13 @@
  */
 
 import React, { useContext, useEffect, useState } from 'react';
-import Box from '@material-ui/core/Box';
-import Typography from '@material-ui/core/Typography';
-import Dialog from '@material-ui/core/Dialog';
-import DialogContent from '@material-ui/core/DialogContent';
-import DialogContentText from '@material-ui/core/DialogContentText';
-import Icon from '@material-ui/core/Icon';
-import IconButton from '@material-ui/core/IconButton';
 import Alert from 'AppComponents/Shared/Alert';
 import { Progress } from 'AppComponents/Shared';
 import API from 'AppData/api';
 import ApiContext from 'AppComponents/Apis/Details/components/ApiContext';
+import ViewPolicyShared from 'AppComponents/Shared/PoliciesUI/ViewPolicy';
 import type { Policy, PolicySpec } from './Types';
 import PolicyViewForm from './PolicyForm/PolicyViewForm';
-
-/** Shared UI Component */
-interface ViewPolicySharedProps {
-    handleDialogClose: () => void;
-    dialogOpen: boolean;
-    policyObj: Policy;
-    isLocalToAPI: boolean;
-    stopPropagation: (e: React.MouseEvent<HTMLDivElement, MouseEvent>) => void;
-    toggleOpen: () => void;
-    policySpec: PolicySpec;
-}
-
-const ViewPolicyShared: React.FC<ViewPolicySharedProps> = ({
-    handleDialogClose,
-    dialogOpen,
-    policyObj,
-    stopPropagation,
-    toggleOpen,
-    policySpec,
-}) => {
-    return (
-        <>
-            <Dialog
-                maxWidth='md'
-                open={dialogOpen}
-                aria-labelledby='form-dialog-title'
-                onClose={handleDialogClose}
-                onClick={stopPropagation}
-                fullWidth
-            >
-                <Box
-                    display='flex'
-                    justifyContent='space-between'
-                    alignItems='center'
-                    flexDirection='row'
-                    px={3}
-                    pt={3}
-                >
-                    <Box display='flex'>
-                        <Typography variant='h4' component='h2'>
-                            {policyObj.displayName}
-                        </Typography>
-                    </Box>
-                    <Box display='flex'>
-                        <IconButton
-                            color='inherit'
-                            onClick={toggleOpen}
-                            aria-label='Close'
-                        >
-                            <Icon>close</Icon>
-                        </IconButton>
-                    </Box>
-                </Box>
-                <DialogContent>
-                    <Box my={2}>
-                        <DialogContentText>
-                            <PolicyViewForm
-                                policySpec={policySpec}
-                                onDone={toggleOpen}
-                            />
-                        </DialogContentText>
-                    </Box>
-                </DialogContent>
-            </Dialog>
-        </>
-    );
-};
-/** Shared UI Component ends */
 
 interface ViewPolicyProps {
     handleDialogClose: () => void;
@@ -192,6 +118,7 @@ const ViewPolicy: React.FC<ViewPolicyProps> = ({
             stopPropagation={stopPropagation}
             toggleOpen={toggleOpen}
             policySpec={policySpec}
+            PolicyViewForm={PolicyViewForm}
         />
     );
 };
